Data format fixed. File XML fixed. Manifest Updated.

+++ src/main/java/inputParser/Product.java
@@ -1,10 +1,13 @@
 package inputParser;
 
+
+import java.math.BigDecimal;
+
 public class Product {
     private String name;
-    private double price;
+    private BigDecimal price;
 
-    public Product(String name, double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -13,7 +16,7 @@ public class Product {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
